requirements.txt
Flask
werkzeug
flask_sqlalchemy
sqlalchemy
flask_login
python-dotenv
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.8.3
reportlab==4.1.0
